# Beautification + String: corrected usprintf return type to size_t

+++ Inc/Common/string.h
@@ -168,7 +168,7 @@ uint8_t STRING_Splitter(char *source, char delimiterChar, char **separated, uint
 size_t string_printf(char *str, const char *format, va_list ap);
 size_t string_printf_safe(char *str, size_t maxLen, const char *format, va_list ap);
 
-uint8_t usprintf(char *str, const char *format, ...);
+size_t usprintf(char *str, const char *format, ...);
 size_t usnprintf(char * str, size_t maxLen, const char * format, ...);
 
 // Unit test
